requirements.txt
pandas
numpy
MetaTrader5
scikit-learn
python-telegram-bot
joblib
